% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{downloadClipDataset_mirwalk}
\alias{downloadClipDataset_mirwalk}
\title{Download CLIP datasets from mirwalk}
\usage{
downloadClipDataset_mirwalk(species = "hsa",
  baseurl = "http://zmf.umm.uni-heidelberg.de/apps/zmf/mirwalk2/downloads/clip/",
  basedir = basedir)
}
\arguments{
\item{baseurl}{(default from website)}

\item{basedir}{- directory where the data should be stored}
}
\value{
TRUE if finished correctly
}
\description{
This function bulk downloads all CLIP-Seq data for miRNAs of a given species
}

