% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{downloadLncInteraction_mirwalk}
\alias{downloadLncInteraction_mirwalk}
\title{Download lncRNA interaction datasets from mirwalk}
\usage{
downloadLncInteraction_mirwalk(species = "hsa",
  baseurl = "http://zmf.umm.uni-heidelberg.de/apps/zmf/mirwalk2/downloads/lncrna/",
  basedir = basedir)
}
\arguments{
\item{baseurl}{(default from website)}

\item{basedir}{- directory where the data should be stored}
}
\value{
TRUE if finished correctly
}
\description{
This function bulk downloads all lncRNA interaction data for miRNAs of a given species
}

