% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{downloadPathwayInteractions_mirwalk}
\alias{downloadPathwayInteractions_mirwalk}
\title{Download pathway - miRNA - datasets}
\usage{
downloadPathwayInteractions_mirwalk(species = "hsa",
  baseurl = "http://zmf.umm.uni-heidelberg.de/apps/zmf/mirwalk2/downloads/vtm/",
  basedir = basedir)
}
\arguments{
\item{baseurl}{(default from website)}

\item{basedir}{- directory where the data should be stored}
}
\value{
TRUE if finished correctly
}
\description{
This function bulk downloads all pathway datasets for miRNAs of a given species
}

