% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{downloadPredictedTargets_mirwalk}
\alias{downloadPredictedTargets_mirwalk}
\title{Download target Predictions}
\usage{
downloadPredictedTargets_mirwalk(species = species,
  baseurl = "http://zmf.umm.uni-heidelberg.de/apps/zmf/mirwalk2/downloads/r/",
  basedir = basedir)
}
\arguments{
\item{baseurl}{(default from website)}

\item{basedir}{- directory where the data should be stored}
}
\value{
TRUE if finished correctly
}
\description{
This function bulk downloads all predicted miRNA targets of a given species
}

