% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{fetchClipDataset_mirwalk}
\alias{fetchClipDataset_mirwalk}
\title{Downloads the object from mirwalk}
\usage{
fetchClipDataset_mirwalk(species = "hsa", geneRegion = "utr5",
  method = "CLASH", identifier = "refseq",
  baseurl = "http://zmf.umm.uni-heidelberg.de/apps/zmf/mirwalk2/downloads/clip/")
}
\arguments{
\item{species}{("hsa","mmu","rno")}

\item{geneRegion}{("utr5","cds","utr3")}

\item{method}{- (mmu: PARCLIP","HITCLIP","CLIPSEQ", hsa: "CLASH","PARCLIP","HITCLIP","iCLIP","CLIPSEQ")}

\item{identifier}{- (refseq)}

\item{baseurl}{- base url of the resource}
}
\value{
TRUE if finished correctly
}
\description{
This function downloads CLIP datasets from mirwalk and directly loads the object into workspace
}

