% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{fetchPredictedTargets_mirwalk}
\alias{fetchPredictedTargets_mirwalk}
\title{Downloads the object from mirwalk}
\usage{
fetchPredictedTargets_mirwalk(species = "hsa", geneRegion = "utr5",
  identifier = "gene_symbol",
  baseurl = "http://zmf.umm.uni-heidelberg.de/apps/zmf/mirwalk2/downloads/r/")
}
\arguments{
\item{species}{("hsa","mmu","rno")}

\item{geneRegion}{("utr5","cds","utr3", "promoter")}

\item{identifier}{- (gene_symbol, entrez)}

\item{baseurl}{- base url of the resource}
}
\value{
TRUE if finished correctly
}
\description{
This function downloads predicted miRNA targets from mirwalk and directly loads the object into workspace
}

