% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{getClipDataset_mirwalk}
\alias{getClipDataset_mirwalk}
\title{Loads a stored object into workspace}
\usage{
getClipDataset_mirwalk(species = "hsa", geneRegion = "utr5",
  method = "CLASH", identifier = "refseq", basedir)
}
\arguments{
\item{species}{("hsa","mmu")}

\item{geneRegion}{("utr5","cds","utr3")}

\item{method}{- (mmu: PARCLIP","HITCLIP","CLIPSEQ", hsa: "CLASH","PARCLIP","HITCLIP","iCLIP","CLIPSEQ")}

\item{identifier}{- (refseq)}

\item{basedir}{- the base directory where the data is stored}
}
\value{
TRUE if finished correctly
}
\description{
This function loads the CLIP data into workspace
}

