% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{getLncInteraction_mirwalk}
\alias{getLncInteraction_mirwalk}
\title{Loads a stored object into workspace}
\usage{
getLncInteraction_mirwalk(species = "hsa", identifier = "gene_symbol",
  basedir)
}
\arguments{
\item{species}{("hsa","mmu","rno")}

\item{identifier}{- ("gene_symbol","ensembl", rno:"noncode")}

\item{basedir}{- the base directory where the data is stored}
}
\value{
TRUE if finished correctly
}
\description{
This function loads the CLIP data into workspace
}

