% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MirWalkRetrieval.R
\name{getPathwayInteractions_mirwalk}
\alias{getPathwayInteractions_mirwalk}
\title{Loads a stored object into workspace}
\usage{
getPathwayInteractions_mirwalk(species = "hsa", identifier = "kegg",
  basedir)
}
\arguments{
\item{species}{("hsa","mmu","rno")}

\item{identifier}{- ("kegg","wiki","panther","GOBP","GOMF","GOCC","gClass","pClass")}

\item{basedir}{- the base directory where the data is stored}
}
\value{
TRUE if finished correctly
}
\description{
This function loads the pathway interactions into workspace
}

